pyproject.toml
[tool.poetry]
name = "apibankdio"
version = "0.1.0"
description = "Api Desafio DIO"
authors = ["Rodrigo Oliveira <rodrigo-fsa@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
pyjwt = "^2.10.1"
uvicorn = "^0.32.1"
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
